Added JumpCounter scene to count jumps
Jump Counter scene is added to the rock climbing minigame scene so the game can display the amount of jumps the player has left.

I will be adding a check to only display the counter when the minigame is being played.

Also, the jumps should be reset whenever the minigame is started again.